Use a markdown table in the man page and convert it on the fly
Instead of duplicating the content of the table in two formats it is now
formatted as a markdown table.  This format is relatively easy to parse
and translate into an scdoc table or a list depending on the LEGACY_ROFF
variable.

This has the nice side effect that the table now renders nicely on
GitHub.